requirements.txt
Django==2.2.4
gunicorn==19.9.0
psycopg2==2.8.3 --no-binary psycopg2
django-suit==0.2.26
redis==3.3.7
django-celery-beat==1.5.0
celery==4.3.0
pytz
flower==0.9.3
djangorestframework==3.10.3
requests==2.22.0
slackclient==2.2.0
# ask for `hipo-django-core`

google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0